use float for weight

+++ Contractor/EdgeBasedGraphFactory.cpp
@@ -70,7 +70,7 @@ EdgeBasedGraphFactory::EdgeBasedGraphFactory(int nodes, std::vector<NodeBasedEdg
         if(edge.source == edge.target)
             continue;
 
-        edge.data.speed = (int)i->speed();
+        edge.data.speed = i->speed();
 		edge.data.distance = (std::max)((int)i->weight(), 1 );
         //INFO( "distance: " << edge.data.distance );
         //INFO( "speed: " << edge.data.speed );
